Added the rect bounding box for resource objects
Also added unit tests for the the bounding box as well. We had to do a bit
of math to calculate the correct bounding box since the Y values appeared
to be flipped.

Another thing to note is that there seems to be a bit of a padding around
the bounding boxs--2 pixels to be exact. This may be an issue hit areas
are grouped too closely together.
to be exact.